# internal/api/spec/api-spec.yaml
openapi: 3.0.0
info:
  title: Hyperboard API
  version: 0.1.0
paths:
...
  /api/v1/tagCategories:
    get:
      operationId: getTagCategories
      summary: List tag categories.
      parameters:
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          $ref: "#/components/responses/TagCategoriesResponse"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "429":
          $ref: "#/components/responses/TooManyRequestsResponse"
        "500":
          $ref: "#/components/responses/InternalServerErrorResponse"
  /api/v1/tagCategories/{tagCategory}:
    put:
      operationId: putTagCategory
      summary: Create or replace a tag category by name.
      parameters:
        - $ref: "#/components/parameters/tagCategory"
      requestBody:
        $ref: "#/components/requestBodies/TagCategoryRequest"
      responses:
        "200":
          $ref: "#/components/responses/TagCategoryResponse"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        "429":
          $ref: "#/components/responses/TooManyRequestsResponse"
        "500":
          $ref: "#/components/responses/InternalServerErrorResponse"
    get:
      operationId: getTagCategory
      summary: Get a tag category by name.
      parameters:
        - $ref: "#/components/parameters/tagCategory"
      responses:
        "200":
          $ref: "#/components/responses/TagCategoryResponse"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        "429":
          $ref: "#/components/responses/TooManyRequestsResponse"
        "500":
          $ref: "#/components/responses/InternalServerErrorResponse"
    delete:
      operationId: deleteTagCategory
      summary: Delete a tag category. All tags with this category will have their category unset.
      parameters:
        - $ref: "#/components/parameters/tagCategory"
      responses:
        "204":
          $ref: "#/components/responses/DeletedResponse"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        "429":
          $ref: "#/components/responses/TooManyRequestsResponse"
        "500":
          $ref: "#/components/responses/InternalServerErrorResponse"
...
  /api/v1/posts/{id}/similar:
    get:
      operationId: getSimilarPosts
      summary: Get posts that are visually similar to the given post based on perceptual hash.
      parameters:
        - $ref: "#/components/parameters/id"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          $ref: "#/components/responses/PostsResponse"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        "429":
          $ref: "#/components/responses/TooManyRequestsResponse"
        "500":
          $ref: "#/components/responses/InternalServerErrorResponse"
...
components:
  schemas:
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    Cursor:
      type: string
    Limit:
      type: integer
      minimum: 1
      maximum: 1024
      default: 24
...
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        $ref: "types-spec.yaml#/components/schemas/ID"
    tagCategory:
      name: tagCategory
      in: path
      required: true
      schema:
        $ref: "types-spec.yaml#/components/schemas/TagCategoryName"
...
    cursor:
      name: cursor
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/Cursor"
    limit:
      name: limit
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/Limit"
...
  requestBodies:
...
    TagCategoryRequest:
      required: true
      content:
        application/json:
          schema:
            $ref: "types-spec.yaml#/components/schemas/TagCategory"
...
  responses:
...
    DeletedResponse:
      description: Resource deleted successfully
    BadRequestResponse:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnauthorizedResponse:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ForbiddenResponse:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFoundResponse:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    TooManyRequestsResponse:
      headers:
        Retry-After:
          schema:
            type: integer
            minimum: 0
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalServerErrorResponse:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    TagCategoryResponse:
      content:
        application/json:
          schema:
            $ref: "types-spec.yaml#/components/schemas/TagCategory"
    TagCategoriesResponse:
      content:
        application/json:
          schema:
            type: object
            properties:
              items:
                type: array
                items:
                  $ref: "types-spec.yaml#/components/schemas/TagCategory"
              cursor:
                $ref: "#/components/schemas/Cursor"
...
    PostsResponse:
      content:
        application/json:
          schema:
            type: object
            properties:
              items:
                type: array
                items:
                  $ref: "types-spec.yaml#/components/schemas/Post"
              cursor:
                $ref: "#/components/schemas/Cursor"
...